refactor: reorganize files in commons package
