+++ commons/src/index.ts
@@ -4,40 +4,11 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-export * from './message-transporters/mocked-backend-message-transporter.js'
-export * from './message-transporters/message.js'
-export * from './message-transporters/message-transporter.js'
-export * from './message-transporters/realtime-user.js'
-export * from './message-transporters/websocket-transporter.js'
-
-export { parseUrl } from './utils/parse-url.js'
-export {
-  MissingTrailingSlashError,
-  WrongProtocolError
-} from './utils/errors.js'
-export * from './utils/permissions.js'
-export * from './utils/permissions.types.js'
-
-export * from './y-doc-sync/y-doc-sync-client-adapter.js'
-export * from './y-doc-sync/y-doc-sync-server-adapter.js'
-export * from './y-doc-sync/y-doc-sync-adapter.js'
-export * from './y-doc-sync/realtime-doc.js'
-
-export { waitForOtherPromisesToFinish } from './utils/wait-for-other-promises-to-finish.js'
-
-export * from './frontmatter-extractor/extractor.js'
-export * from './frontmatter-extractor/types.js'
-
-export * from './note-frontmatter-parser/parse-raw-frontmatter-from-yaml.js'
-export * from './note-frontmatter-parser/convert-raw-frontmatter-to-note-frontmatter.js'
-export * from './note-frontmatter-parser/default-values.js'
-export * from './note-frontmatter-parser/parse-tags.js'
-export * from './note-frontmatter-parser/types.js'
-
-export * from './note-frontmatter/iso6391.js'
-export * from './note-frontmatter/frontmatter.js'
-export * from './note-frontmatter/slide-show-options.js'
-
-export * from './title-extraction/generate-note-title.js'
-
-export { extractFirstHeading } from './title-extraction/extract-first-heading.js'
+export * from './frontmatter-extractor/index.js'
+export * from './message-transporters/index.js'
+export * from './note-frontmatter/index.js'
+export * from './note-frontmatter-parser/index.js'
+export * from './parse-url/index.js'
+export * from './permissions/index.js'
+export * from './title-extraction/index.js'
+export * from './y-doc-sync/index.js'
