+++ commons/src/permissions/index.ts
@@ -1,0 +1,8 @@
+/*
+ * SPDX-FileCopyrightText: 2023 The HedgeDoc developers (see AUTHORS file)
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ */
+
+export * from './permissions.js'
+export * from './permissions.types.js'
