+++ commons/src/title-extraction/index.ts
@@ -1,0 +1,8 @@
+/*
+ * SPDX-FileCopyrightText: 2023 The HedgeDoc developers (see AUTHORS file)
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ */
+
+export * from './generate-note-title.js'
+export * from './extract-first-heading.js'
