+++ commons/src/y-doc-sync/index.ts
@@ -1,0 +1,9 @@
+/*
+ * SPDX-FileCopyrightText: 2023 The HedgeDoc developers (see AUTHORS file)
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ */
+export * from './y-doc-sync-client-adapter.js'
+export * from './y-doc-sync-server-adapter.js'
+export * from './y-doc-sync-adapter.js'
+export * from './realtime-doc.js'
